Refactors admin functionality into specific controllers
Removes AdminController and splits its responsibilities into ConferenceController and PageController for better modularity and maintainability. Updates API routes to reflect the new structure.

Improves code organization and aligns with single-responsibility principles.
